First fix for Plasma 6.5 compatible

+++ update.sh
@@ -2,6 +2,13 @@
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 sh $SCRIPT_DIR/build.sh
 kpackagetool6 -u $SCRIPT_DIR/release/FancyTasks.tar.gz
-killall plasmashell
-kstart5 plasmashell
+
+# Restart plasmashell
+if systemctl --user is-active --quiet plasma-plasmashell.service; then
+    systemctl --user restart plasma-plasmashell.service
+else
+    killall plasmashell
+    plasmashell > /dev/null 2>&1 &
+fi
+
 sh ./iconinstall.sh
